+++ vitrina/resources/views.py
@@ -45,6 +45,12 @@ class ResourceDetailView(PermissionRequiredMixin, HistoryMixin, DatasetStructure
     def get_detail_url(self):
         return self.object.dataset.get_absolute_url()
 
+    def get_child_resources_url(self) -> str:
+        return reverse(
+            "dataset-child-resources",
+            args=[self.object.dataset.pk],
+        )
+
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         context["resource"] = self.object